refactor: split monolithic errand components
Split SupportErrandComponent (202 -> 13 lines) and
CasedataErrandComponent (285 -> 13 lines) into focused
sub-components: data hooks, form providers, headers,
layouts and utility functions.
